requirements.txt
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
six==1.14.0
Werkzeug==1.0.0
WTForms==2.3.1
